+++ NPCMapLocations/docs/release-notes.md
@@ -1,6 +1,16 @@
 ﻿﻿﻿[← back to readme](README.md)
 
 # Release notes
+## 2.7.1
+Released 26 November 2021 for SMAPI 3.12.5. Updated by Pathoschild.
+
+* Improved the `npc_map_locations_summary` command:
+  * Added a table of location data.
+  * The map vectors list now includes those bundled with NPC Map Locations.
+* Fixed farm buildings not shown on load if any location without map vectors is also excluded.
+* Fixed Caldera not handled as part of the Volcano Dungeon.
+* Improved translations. Thanks to Zangorr (added Polish)!
+
 ## 2.7.0
 Released 20 October 2021 for SMAPI 3.12.5. Updated by Pathoschild.